+++ src/renderer/state/preferences/app-settings.ts
@@ -150,10 +150,6 @@ export const deleteBranchOnArchiveAtom = settingAtom(
 	'git',
 	'deleteLocalBranchOnArchive',
 );
-export const reclaimDiskOnArchiveAtom = settingAtom(
-	'git',
-	'reclaimDiskOnArchive',
-);
 export const archiveOnMergeAtom = settingAtom('git', 'archiveAfterMerge');
 export const setUpstreamOnPushAtom = settingAtom('git', 'setUpstreamOnPush');